requirements.txt
#bcrypt==4.0.1
dotenv==0.9.9
fastapi==0.116.1
httpx==0.28.1
passlib==1.7.4
PyJWT==2.10.1
pydantic==2.11.7
pydantic[email]==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pytest-asyncio==0.21.1
passlib[bcrypt]==1.7.4
requests==2.32.5
SQLAlchemy==2.0.43
#uvicorn==0.35.0
uvicorn[standard]==0.35.0
